parsing for var exp to save almost done.  Need to look into why my parse for quotes arr[1] is not working;
