+++ src/parsing/delimiter.c
@@ -1,0 +1,62 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   delimiter_expansions.c                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kbolon <kbolon@42.fr>                      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/04/18 14:38:21 by kbolon            #+#    #+#             */
+/*   Updated: 2024/05/01 11:59:03 by kbolon           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../minishell.h"
+
+int	find_delimiter(char *s1, char *delim)
+{
+	int	i;
+	int	j;
+
+	i = 0;
+	j = 0;
+	printf("in find delimiter \n");
+	while (s1[i] != '\0' && delim[j] != '\0' && s1[i] == delim[j])
+	{
+		i++;
+		j++;
+	}
+	if (delim[j] == '\0')
+		return (1);
+	printf("out find delimiter \n");
+	return (0);
+}
+
+char	*check_quotes(char *s)
+{
+	size_t		i;
+	size_t		len;
+	char		*temp;
+
+	i = 0;
+	len = ft_strlen(s);
+	temp = s;
+	if ((s[0] == '\'' && s[len - 1] == '\'') || (s[0] == '\"' && s[len - 1] == '\"'))
+	{
+		if (s[i] == s[len - 1])
+		{
+			temp = (char *)malloc(sizeof(char) * len -2);
+			if (!temp)
+			{
+				printf("problems mem alloc checkquotes");
+				exit (1);
+			}
+			while (i < len - 2)
+			{
+				temp[i] = s[i + 1];
+				i++;
+			}
+			temp[i] = '\0';
+		}
+	}
+	return (temp);
+}
