+++ src/parsing/parse_echo_awk.c
@@ -6,7 +6,7 @@
 /*   By: kbolon <kbolon@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/15 12:21:12 by kbolon            #+#    #+#             */
-/*   Updated: 2024/04/26 18:59:16 by kbolon           ###   ########.fr       */
+/*   Updated: 2024/05/01 11:52:27 by kbolon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
